Updated from testing, work in progress.
Popping back a syntax level needs to be updated.  It currently updates the
current syntax tree frame to the SUB entry for the level being popped,
when there is something to save.  However, that can mess up the type of
syntax tree entry pointed to by that stack frame.

Need to create a special stack frame just for the SUB entry, or separate
pointers in stack frames to the tree entry corresponding to that frame,
and the current entry to add to.