Parse stack entries now have single pointer to syn tree entry.
Each parse stack entry now has a single pointer to the syntax tree entry
that was current when that parse stack entry was created.  This allows
popping state properly.  Was previously using the current syn tree entry
pointer, but that can change over the lifetime of a parse stack entry.